+++ src/main/java/dev/gaycookie/geodes/GeodeLootTable.java
@@ -5,21 +5,25 @@ import java.util.List;
 import org.apache.commons.lang3.RandomUtils;
 
 public class GeodeLootTable {
-  private List<Object> loot_table;
+  private List<Object> loot_table_overworld;
+  private List<Object> loot_table_nether;
+  private List<Object> loot_table_end;
   private GeodeLootTableEntry last_item;
 
-  public GeodeLootTable(List<Object> loot_table) {
-    this.loot_table = loot_table;
+  public GeodeLootTable(ConfigClass config) {
+    this.loot_table_overworld = config.overworld_loot_table;
+    this.loot_table_nether = config.nether_loot_table;
+    this.loot_table_end = config.end_loot_table;
   }
 
-  public GeodeLootTableEntry getRandomEntry() {
+  public GeodeLootTableEntry getRandomOverworldEntry() {
     double random = RandomUtils.nextDouble(0.0, 100.0);
 
-    for (Object geodeTableEntry : this.loot_table) {
+    for (Object geodeTableEntry : this.loot_table_overworld) {
       GeodeLootTableEntry entry = (GeodeLootTableEntry) geodeTableEntry;
 
       if (random <= entry.weight) {
-        if (entry == this.last_item) return this.getRandomEntry();
+        if (entry == this.last_item) return this.getRandomOverworldEntry();
         
         if (entry.item != "minecraft:air") last_item = entry;
         return entry;
@@ -29,7 +33,31 @@ public class GeodeLootTable {
     return null;
   }
 
-  public List<Object> getLootTable() {
-    return this.loot_table;
+  public GeodeLootTableEntry getRandomNetherEntry() {
+    double random = RandomUtils.nextDouble(0.0, 100.0);
+
+    for (Object geodeTableEntry : this.loot_table_nether) {
+      GeodeLootTableEntry entry = (GeodeLootTableEntry) geodeTableEntry;
+
+      if (random <= entry.weight) {
+        return entry;
+      }
+    }
+
+    return null;
+  }
+
+  public GeodeLootTableEntry getRandomEndEntry() {
+    double random = RandomUtils.nextDouble(0.0, 100.0);
+
+    for (Object geodeTableEntry : this.loot_table_end) {
+      GeodeLootTableEntry entry = (GeodeLootTableEntry) geodeTableEntry;
+
+      if (random <= entry.weight) {
+        return entry;
+      }
+    }
+
+    return null;
   }
 }