+++ src/main/java/dev/gaycookie/geodes/items/EndStoneGeode.java
@@ -1,0 +1,7 @@
+package dev.gaycookie.geodes.items;
+
+public class EndStoneGeode extends GeodeItem {
+    public EndStoneGeode() {
+        super();
+    }
+}