+++ src/main/java/dev/gaycookie/geodes/items/NetherrackGeode.java
@@ -1,0 +1,7 @@
+package dev.gaycookie.geodes.items;
+
+public class NetherrackGeode extends GeodeItem {
+    public NetherrackGeode() {
+        super();
+    }
+}